backend/requirements.txt
Django>=4.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.6,<5.0
psycopg[binary]>=3.2,<4.0
pgvector>=0.3,<1.0
httpx>=0.27,<1.0
tenacity>=8.2,<9.0
